+++ factory/crypto/signatureHolder.go
@@ -1,4 +1,4 @@
-package signing
+package crypto
 
 import (
 	"sync"
